CameraControl
A short script to let the camera rotate around the earth when right-clicking and dragging the mouse. Also, using the mouse scrollwheel changes the "zoom" of the camera (well, it's changing it's position, but w/e).

PS_1 : Added a start of a script that'll lead to the clickable-earth to select a country.
PS_2 : Unity got updated, i'm on 2017.3.0.0f3. Get up-to-date you sluts !
PS_3 : Started using Rider instead of Monodevelop to code (personnal)